requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic 
redis
numpy
sentence-transformers